+++ webapp/src/css/footer.css
@@ -7,5 +7,5 @@
 }
 
 #link {
-    color: #000000;
+    color: #0062fb;
 }
