+++ webapp/src/css/home.css
@@ -7,7 +7,9 @@
     margin-top: 20px;
 }
 
-#screen {
+#homeScreen {
+    display: flex;
+    flex-direction: column;
     text-align: center;
 }
